Remove DSv4 sparse cache utils split
